chore: apply review from meng
